refactor: runs tests against transpiled version
This is a temporary solution to validate transpiled code, and after
conversion it should run tests against TypeScript code.
